requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.1.0
streamlit>=1.28.0
plotly>=5.15.0
seaborn>=0.12.0
matplotlib>=3.6.0
dvc>=2.58.0
joblib>=1.2.0
pyyaml>=6.0
datetime
python-dateutil>=2.8.0
xgboost>=1.7.0
lightgbm>=3.3.0